requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==23.0.0
pydub==0.25.1
google-cloud-speech==2.23.0
requests==2.31.0
pillow==10.1.0
flask-migrate
flask-wtf
google-generativeai
sqlalchemy
wtforms
dotenv
email_validator
psutil
python-dateutil==2.8.2
SpeechRecognition==3.10.0
